requirements.txt
pyrebase4
streamlit
pandas
smtplib
gspread
oauth2client
df2gspread